Move Translator class to a new Services namespace, since more Services will follow
